Remove rankings per as_of_date if as_of_date is in index [Resolves #59]
Computing rankings when there are daily predictions was always of dubious utility, and was done to conform to the Predictions schema at what seemed like a low cost. As it turns out, the code to subset these dataframes was very memory-hungry, so for now the cost for leaving those columns null seems to be lower.

Also, bulk_save_objects is used now because the scale of these predictions can get pretty big.

- Remove ranking code from multiple as_of_dates-in-matrix code branch
- Switch Predictions.write_to_db to bulk_save_objects