Support Parallelization [Resolves #43]
The main goal here is to separate uuid creation and matrix creation, so duplication issues can be resolved before parallelization is begun. Two interfaces to the Architect are introduced: generate_plans, and build_all_matrices. The former is concerned with taking matrix definitions and both augmenting these definitions with uuids and generating a duplicate-free list of arguments for the `build_matrix` method. The latter is a simple loop over build tasks to create matrices. This is meant to an external caller parallelizing matrix creation need only replace the latter method, and not concern itself with the complex business logic of working with uuids and generation of build tasks.

Also needed to support parallelization is the prefixing of the entity date table and feature and label csvs with the matrix uuid, so parallel processes can work without clobbering each other.

- Return table name from make_entity_date_table, now with uuid
- Pass entity_date_table_name and matrix_uuid to write_features_data and write_labels data and use to create unique filenames
- Split functionality from chop_data and design_matrix into generate_plans and build_all_matrices