Add validation and support for restricting the "beginning of time" (#74)
* Add support for restricting the "beginning of time"

Adds a new keyword parameter for SpacetimeAggregations that enables restricting the rows included in their calculations based upon an absolute minimum date.

Adds actual behavior tests for SpacetimeAggregation with testing.postgresql.

This takes a SQL connection to allow validation against a SQL server. By default, `Aggregation.execute()` will call the validate method.  SpacetimeAggregations now raise an error in the case where a date/interval combination happens to cross before the beginning of time (so long as the interval is not all).

If this proves to be annoying, we can perhaps change it to be a warning or even add an optional override to explicitly allow this to occur.  But I think it behooves us to start conservatively.