Filter by Entity State [Resolves #1]
We want to be able to consider entity state when generating matrices; for example, whether or not a building was permitted at the time. The general data flow to include entity states is as follows:

Label Generation and Feature Generation *do not* include any state filtering. This is because we would like to iterate through different state filters within a given experiment. This required removing the existing state filter code from the LabelGenerator

State filtering happens when we create the entity_date table for a given matrix. In fact, the sparse state table is pretty close to what an entity date table should be. Because of this, when we construct the queries to grab labels and features from the database, we don't need to concern ourselves with states as long as we join with this table. At a high level, nothing else really needed to change. A lower level list of changes is below:

- Remove Builder#build_labels_query, replace with _all_labeled_entity_dates_query for train matrices.
- Add state handling to _all_labeled_entity_dates_query() and _all_valid_entity_dates_query()
- Modify Builder#write_labels_data to call _outer_join_query for both train and test matrices, since the branching logic they need is handled when creating the entity-date table.
- Remove as_of_times logic from outer_join_query, since this is handled when created the matrix's entity-date table
- Rename query string-returning methods to pseudo-private methods without the term 'build'.
- Rename tmp_entity_date to matrix_entity_date for clarity, make it into temporary table that doesn't need to be manually cleaned up in build_matrix
- Remove state handling from BinaryLabelGenerator as it is handled on the matrix level.
- Add state handling to Planner
- Move test logic for build_labels_query up one level into a new test for write_labels_data
- PEP8 fixes