Order prediction matrix columns to match train matrix [Resolves #170] (#172)
We want to make sure that when we process a prediction matrix, the columns match up with that of the train matrix. This turned out a bit more complicated than I wanted to, to keep some of the flexibility in triage.Storage (that is being used!).

When the InMemoryMatrixStore is used (example: https://github.com/dssg/police-eis/blob/master/eis/run_models.py#L306 ), it is impossible to infer anything about the train matrix. For that reason, we have to support the passing in of the column list itself. To handle this, we have an if statement in predictor.predict, that uses this when present but otherwise retrieves the column list by retrieving the train uuid, finding the stored matrix, and grabbing the column names from there. I don't love the end result, but I also can't think of a better alternative.

- Add util to retrieve train matrix id given a trained model
- Add method to MatrixStore to return a matrix store of the same class with a different UUID
- Add method to MatrixStore to reorder the matrix's columns
- Add train_matrix_columns optional argument to Predictor#predict, usable when the MatrixStore can't support retrieving the train matrix columns after the fact (e.g. when InMemoryMatrixStore is used)
- Add test utility to create basic metta matrices/metadata
- Give train and test matrices in unit test different column orders
- Handle more cases matrix column sorting scenarios