Rename Pipeline to Experiment (#174)
Triage's scope has narrowed, in a way, by ceasing to implement core components and only becoming the interface. In another way, it's broadened, to where different interfaces will be implemented here. Because of this, the name 'Pipeline' currently used is far too broad: it refers to an Experiment, which is just one phase of a project.

The subclasses prefixes have also been renamed, partly because of this change but also they just were not as descriptive as they could be.

- Renamings: Pipeline->Experiment, SerialPipeline->SingleThreadedExperiment, LocalParallelPipeline->MultiCoreExperiment
- Rewrite README to reflect Triage's new role