Component Validation [Resolves #8]
Components should offer some validation on their inputs before they are forced to do computationally-intensive work. This pull request adds a `validate` method to the components here with meaningful configuration. Generally this has no parameters.

The FeatureGenerator validation is slightly different, as current the config isn't bound at object creation time. I think this is a flaw that we should fix, but didn't think it was suitable to mix that work in with this pull request, so the validation method takes the config as an parameter.

- Add database_reflection module to handle querying of database state
- Add validations module to define common validations
- Add validation methods to StateTableGenerator, LabelGenerator, FeatureGenerator, Planner, Builder, and FeatureGroupCreator
- Add validation calls to integration test to help defend against false positive validations