Fix import in feature_importances [Resolves #13]
If sklearn.linear_model isn't already imported, the line in question will fail even if sklearn is imported at the top of the file. This was hidden because the test has to import sklearn.linear_model in order to create the model we're testing. Although it's very hard to add a test that catches this, this change should fix the bug.