Decouple Planner from Builder [Resolves #269]
Once upon a time, the Planner and Builder classes lived in one Timechop class called 'The Architect'. Eventually, the different functionalities (figuring out and deduping what matrices needed to be built, and then building matrices one-by-one) diverged into the two classes you see today, but to support backwards compatibility Planner delegated the building itself to Builder, as opposed to letting the Experiment have its own reference to a builder. This half-solution has lived on for far too long. We will be making much more changes in the builder soon, and it'll be easier to work on in its proper place.

- Remove builder delegation methods from Planner
- Create matrix builder in Experiment, and call build_matrix/build_all_matrices directly from experiments
- Split test_planner_builders into test_planner and test_builders