Generalize labels interface to query
After internal discussion, it seems that the planned early warning and inspections on-rails label generators are not the direction we want to go. They require too much preprocessing to be used in most cases. The parameterized query is more generally usable, and any transformations needed are transparent in the query. It's easy to implement early warning or inspections on an external layer.

The new label config interface has three keys: a query, a fill-missing behavior (True or False; the absense of the key will leave missing labels as missing), and a name that will go in matrix metadata.

- Replace label generator inheritence with one class that takes a query
- Allow passing of fill-missing behavior for labels in config, to allow early warning style labels [Resolves #281]