Enforce alembic use for experiments, add CLI managements for alembic … (#415)
* Enforce alembic use for experiments, add CLI managements for alembic [Resolves #412]

It's advantageous to force the use of Alembic for schema migrations now. The Experiment here is changed to do so, and to aid management of this we modify the triage CLI script to cover upgrading a db and stamping a db to a specific version. To make the development of this and general usage of Alembic smoother, there are some fixes to Alembic files (env.py, alembic.ini) that cut down on the labryinthian invocations that Triage needed for Alembic commands before.

Alembic enforcement:
- Replace catwalk ensure_db that uses Base.metadata.create_all with results_schema upgrade_db that uses alembic upgrade head
- Add migration that adds model group stored procedure so alembic upgrade covers it

Usability:
- Add __init__.py scripts to alembic directories so they make it into the installed python package
- Add %(here)s to alembic.ini scripts directory to make it more resilient to being run from different directories
- Modify alembic env.py to add option of more standard config attribute to hold the URL
- Update results_schema README with details about all this