Refactoring/Updating postmodeling
This code creates new ways to set-up the postmodeling analysis. The new
workflow will rely on a model class that will contain all the auditioned
models_ids and its attributes to calculate and visualize different
insights for each of the models. The following is a list of possible
add-ons to the analysis:
 1. Feature importances (sklearn/others)
 2. Rank correlation between features and top-k (can be also from the tail)
 3. Leave one-out / CV for different groups of variables
 4. Error analysis.

Some utils are added for development purposes and for file management of
s3 (when matrices are stored there)