Add new methods to model group postmodeling
This new commit adds functionality to the ModelGroup class. Now, besides
Jaccard similarity, the method can calculate rank correlations for
entities and features. This methods can be applied to all the
model_group_ids selected by the Audition module, or to a subset of those
models (v.gr. apply to only to the latest as_of_date).

The Jupyter notebook is updated to reflect the new functions and plots.

Bugs:
 - Color and other plot arguments are hardcoded.
 - Probably there must be a better/more efficient way to get this done.

To do:
 - Add more plotting arguments into the class method.
 - Fix call of train matrices