Add time comparisons to Jaccard sim plot
This commit fixes to problems with the Jaccard Similarity plot. First,
the plot allowed to compare different models (model_ids) with different
temporal configurations, which is not always desirable when trying to
compare models for a certain prediction in mind. Now the plot_jaccard
method includes a temporal_comparison options which creates a plot for
each date in the predictions table for the selected models.

Second, the filtering for top_n was mistakenly including all
observations because was using score as a variable to select rather than
the rank. This is solved and shown on the Ipython Notebook