Use session.merge for matrix build [Resolves #453]
session.add was used for populating Matrix rows in the database, which
errors if there is already an existing row. This was uncaught before
because there was no 'rerun matrix build with replace=True' test, only a
'rerun matrix build with replace=False' test, which passes because it
does not get to the point of trying to insert the Matrix again.

- Add test to architect.builders that reruns a matrix build with
replace=True to make sure that it runs to completion
- Change session.add to sessoin.merge in architect.builders.build_matrix