Labeler and modeler error methods
This commit separates the error labeling from the error modeling by
creating two different methods. First, the labeler will define the
correct predicted entities using a user threshold, and label the FPR and
FNR. This function returns two objects, one interator with labels, and
the subsetted feature matrix.

Once created by the labeler, the matrix and the outcome iterator are
passed into the modeler to use a RandomTrees model to get analize the
relation between errors and features