Downcast matrices [Resolves #372]
Introduces triage.util.pandas.downcast_matrix, which downcasts the
values of a dataframe to their minimal values (e.g. float64->float32, or
int64->int32). This is called in two places:

1. Within MatrixBuilder to each smaller
matrix before they are joined together, with the intention of stopping
memory from spiking at the join step.

2. When loading a matrix into memory from the MatrixStore class.
Since it made sense to put this in the superclass as opposed to
forcing each subclass to implement it, it was added to the .matrix
getter. While doing this, it made sense to do the same for the set_index
call as well, allowing some further cleaning up of the MatrixStore
subclasses.