Add iterator to threshold parameters
Now the postmodeling config file is able to catch a whole dict of
parameters with different kind of threshold definitions, as rank_abs, or
rank_pct. The Parameters object will take this dictionary and convert it
into an iterator that can be passed to all the functions that require a
threshold and repeat the analysis for all the listed thresholds and
metrics. This will save fingers to the user.