Threshold iterator for error models!
This commit adds a new method to make error analysis per each model_id.
This method takes a ThresholdIterator object and uses this to run the
modeling function for each of the defined thresholds and saves a tree
plot in pdf. If no iterator is passed, you can pass a threshold type and
a threshold.