Add baseline for metric comparison against models
Now a plot for precision or recall at a defined threshold is plotted.
This plot can be done individually for the selected model_group_id's, or
can include a defined baseline in the configuration file (a sql query).
A try clase is used to catch possible errors in the thresholds (i.e.
metric doesn't exists in the evaluations table).