Crosstabs plot added
A new method plots the ratio between means for each feature of the
model. This is the best way to visualize the crosstabs outcome. One nice
thing to have in this code is the ability of running crosstabs from the
notebook, rather than doing it separately. Later we can check that