Change Jaccard Matrix temporal behavior
Now the Jaccard matrix can be plotted not using the unique date in the
prediction matrix, but using the year only. That way is convenient to
plot, and is less error prone when cohorts are defined in a more
complex way. Also, some changes in SQL code are made to add new
infomartion about thresholds