Revamp Ranked Correlation and Jaccard
This commit includes changes in the self.plot_jaccard and
self.plot_ranked_corrlelation_preds allowing the user to use a different
function between features and predictions. Also, the self._rank_corr_df
is more robust to changes in metrics, and will take correctly the
'rank_pct' threshold values.

New auxiliary tables are added to the class. For instance, a new
self.same_time_models table is added with a table with the array of
model_id with the same as_of_date. This is useful when comparing subsets
of predicted values.