Associate Experiment with all models and matrices [Resolves #411] (#476)
* Associate Experiment with all models and matrices [Resolves #411]

To aid in the lookup of relevant matrices and models to an experiment
afterwards, we want to record whether or not that experiment *used* that
matrix or model, regardless of whether or not it had to *build* it.

We still want to record which experiment built the matrix or model. This
was already being recorded in the models.experiment_hash column, which
is renamed to built_by_experiment. The matrices table gains a similar
column to mirror this functionality.

Audition is among the use cases for this change, so the pre-audition query is
updated to reflect this.

- Create experiment_matrices and experiment_models tables
- Rename models.experiment_hash to models.built_by_experiment to contain
what is currently stored there but with a more specific name
- Add matrices.built_by_experiment to mirror the models column.
- Add utils to record these rows and trigger them from the Experiment
experiment_models
- Update preaudition query to look at experiment_models
fixing
- Add note on these new tables to Experiment Algorithm doc and
Experiment Running doc
- Update getcwd patch in audition tests to only exist where it needs to,
fixing a local test failure where the getcwd was screwing up
testing.postgresql