Refactoring: Remove cohort options besides query [Resolves #504]
To simplify the code for some upcoming functional changes, we remove
unused options for cohort configuration. The 'dense_states' and
'entities' options are now gone, with only the favored 'query'
remaining.

We similarly update the name of the generator to 'cohort' for a bit less
confusion, simplify the example experiment config and validation, and remove related old code.