support 5 GB multipart upload threshold via S3Fs
*i.e.* clean-up / speculation for issue #530 and pull request #534

S3Fs *appears* to support the boto3 configuration necessary to set the
multipart upload threshold, (from the API documentation and from a
manual test I've just done).

Moreover, S3Fs appears to provide us a slightly nicer interface, and more
options for performance optimizations, than we could easily get on our
own, with so few lines of code. (For that matter, even as is, this means
fewer lines of code, and more straight-forward code.)