Fix db engine check in Experiment [Resolves #538] (#539)
* Fix db engine check in Experiment [Resolves #538]

The experiment checks whether or not the passed-in engine is a base sqlalchemy engine, and converts it into the serializable triage engine if so.

However, this check was broken. This PR fixes it to actually duck-type,
as well as moves the check to the MultiCore so we can actually error out
if the wrong engine is passed.