Add additional feature group CV strategy (#518)
* Add additional feature group CV strategy

This commit adds a new feature CV strategy to the architect. This
function allow the experiment to take all the combinations of feature
groups additional to the LOO and LOI strategies. The unit test also is
commited and example experiment config is updated.

To-do:
For computational reasons, it would be a nice-to-have set of tests that
avoid to calculate redundant tests when LOI, LOO, and all are specified
at the same time.