Only Build Features for Cohort [Resolves #513] (#567)
* Only Build Features for Cohort [Resolves #513]

By default, only build/save features for the given cohort. This is to
optimize speed for new users (and visit-based problems) to only what is
necessary.

For other users, the `features_ignore_cohort` flag is introduced. If this is
set to True, the old behavior that builds features independently of the
cohort is utilized.

To make the use of this flag in combination with the replace Flag safe,
the feature table "needs features?" check now sees if there are any
cohort rows not present in the imputed table and rebuilds if anything is
missing.