Implement train/testing priority [Resolves #542]
Train/test tasks are now implemented in batches, based on what results
people should be interested in first:

- Batch 1: Short and important ones, like Decision Trees, Scaled Logistic Regressions,
and baselines. These will be parallelized if using an Experiment
subclass that supports parallelization
- Batch 2: Big ones with n_jobs=-1. This will be run serially no matter
what the Experiment subclass is, because the classifier is expected to
parallelize and adding on another layer is likely to crash the
Experiment.
- Batch 3: All others, parallelized similar to batch 1. These are ones
that might be expected to take a while to complete (Gradient Boosting,
forests without n_jobs=-1) and/or ones less likely to be effective.