Fix MatrixStore memory leak [Resolves #594]
Prior to now, the only way a MatrixStore's internal cache was invalidated was
if it was serialized (e.g. to be sent over the wire for
    multiprocessing). Because of this, if many MatrixStores were in
    scope at the same time without being serialized, their memory would
    add up. This never happened in the wild until recently when the
    Experiment was changed to have all model training tasks be
    assembled at once instead of per-split. Because the assembling of
    training tasks retrieves the label vector in order to do a sanity
    check, the matrix does get loaded and so the memory usage grows and
    grows.

The route taken to fix this is to make caching in the MatrixStore opt-in
as opposed to hardcoded (like before), or opt-out (another possible
path). The MatrixStore won't cache the underlying matrix as a dataframe
unless you are in a `with matrix_store.cache():` context manager block.
This fixes the memory leak because while assembling training tasks, we
don't have caching enabled, but during training/testing, we do.

In the process of making this change, the brittleness of the mechanism
for handling the splitting of labels off from the stored matrix came up
    quite a bit in unit tests. To straighten this out, the interface of
    the MatrixStore is now changed. Instead of the `.matrix` property
    which would return either the matrix with labels or the matrix
    without labels (depending on whether or not `.labels()` had been
    called prior, which would pop the labels, there are different
    properties when a reference to each form of the matrix is needed.
    The `.design_matrix` property (usually what is wanted) returns the matrix without labels,
    the `.matrix_with_labels` property returns the matrix with labels,
    and the `.labels` property returns the label vector. New internal
    methods handle the conversion needed here without having to store
    both versions of the matrix. With this change, the contents of the
    returned matrix should not be state-dependent as they were before.