Postmodel Analysis (#482)
New things are added since #431:

Two main classes are created to help the user to interact with post modeling tasks. `ModelEvaluator` and `ModelGroupEvaluator` contained different methods to read and play with triage `model_group_id`s. Most of these task can be seen in action in the included Jupiter Notebook.

A `PostmodelParameters` class is added to read a `yaml` configuration file with different parameters important to the evaluation. These parameters are outlined in the included Jupiter Notebook.

Crosstabs new yaml interface was moved here too.