Faster train/test task generation
Generating train/test tasks can take hours. One culprit may be slow IO
of the matrices, which are loaded to decide if the job should be
processed. This commit makes the train-tester generate all of the tasks
withoug checking the matrices and skips processing the tasks if the
matrices are no good.