Introduce experiment_runs table, beef up experiments table
At long last, the experiment runs table. It contains a variety of metadata about the experiment run, such as installed libraries, git hash, and number of matrices and models built/skipped/errored. Similarly, the experiments table is augmented with data that doesn't change from run-to-run (e.g. number of time splits, as-of-times, total grid size)

A variety of methods on the Experiment act as 'entrypoints'. The first entrypoint you hit when running an experiment (e.g generate_matrices, or train_and_test_models) gets tagged on the experiment_runs row.

- Add Experiment runs table [Resolves #440] [Resolves #403] and
run-invariant columns to Experiments table
- Add tracking module to wrap updates to the experiment_runs table
- Have experiment call tracking module to save initial information and retrieve a run_id to update with more data later, either itself or through components (e.g. MatrixBuilder, ModelTrainer) that do relevent work
- Have experiment save run-invariant information when first computed