Remove support for entity-only matrix indices [Resolves #477] (#622)
* Remove support for entity-only matrix indices [Resolves #477]

There is a lot of branching code to support single-level entity-id
indices in matrices. Since the Experiment was born many eons ago, this
code hasn't really been used, but the code to support it is still there.
No longer! Some tests that support this are removed, and we also
simplify more of the test matrix creation code to reflect this.

* Some test fixes

* Refactor save_db_objects to use Ohio from review

* Another review tweak

* Some refactoring based on review

* Remove unnecessary numpy import