Add compute best/worst/stochastic for each evaluation [Resolves #292]
Instead of just computing evaluation metrics based on one sort seed,
computes best/worst/stochastic (many sort seeds)

- Add schema migration to create new columns to store the different
versions of each metric, plus number of trials and standard deviation.
The old 'value' is copied to stochastic_value and then dropped.
- Modify sorting and thresholding routines to use numpy.arrays instead
of converting to Python lists
- Update sort_predictions_and_labels to implement best and worst sort in
addition to the random one
- Update catwalk.ModelEvaluator to generate evaluations for the
best/worst sorting for each metric, and do 30 trials for the metrics
which have sufficiently different best/worst. To enable this, the
evaluation is refactored to decouple the flattening of the metric definitions from the actual evaluation computation.
- Update Audition and Postmodeling to look at the 'stochastic value' by
default
- Remove sort_seed from scoring example config as it is no longer used