Add more user database management options to CLI [Resolves #697]
In recent weeks/months, more operations on the results schema have
proven to be things that are useful to 'users' (people who use the
'triage' command), not just 'developers' (people who use the 'manage'
command). These include: stamping to a specific revision, downgrading,
upgrading to a specific revision, and even just viewing the revision history.
Here we allow the `triage db` command to interface with
alembic to do these things.

Furthermore, the old 'stamp' logic in `triage db` isn't terribly useful
now that we have been on alembic for a while, and pinning it to
experiment config versions wasn't very useful. Using the standard
alembic revisions for stamping I think makes more sense, but I copied
the dictionary from before into the help text for 'stamp' because it
could still be helpful.

- Modify old `triage db stamp` logic to use standard alembic revisions
- Enable `triage db upgrade` to take a revision (but default to HEAD)
- Add `triage db downgrade` that takes a revision
- Add `triage db history` to show revisions